requirements.txt
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
requests==2.28.2
streamlit==1.18.1
streamlit_lottie==0.0.3
